Refactor file tree organizations
