--- src/ProtectedRoute.jsx
@@ -1,15 +1,0 @@
-import React from "react";
-import { Navigate } from "react-router-dom";
-import { useSelector } from "react-redux";
-
-const ProtectedRoute = ({ redirectPath = "/login", children }) => {
-  const authUser = useSelector((state) => state.authUser.value);
-
-  if (!authUser) {
-    return <Navigate to={redirectPath} replace />;
-  }
-
-  return children;
-};
-
-export default ProtectedRoute;
