+++ src/components/atoms/PollCard.jsx
