+++ src/index.js
@@ -1,11 +1,11 @@
 import React from "react";
-import { createRoot } from "react-dom/client";
+import { store } from "./redux/store";
 import { Provider } from "react-redux";
-import { store } from "./app/store";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
-import "./index.css";
+import App from "./app/App";
 import { BrowserRouter } from "react-router-dom";
+import reportWebVitals from "../src/utils/reportWebVitals";
+import { createRoot } from "react-dom/client";
+import "./css/index.css";
 
 const container = document.getElementById("root");
 const root = createRoot(container);
